requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==25.1.0
    # via cloud-sql-python-connector
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via cloud-sql-python-connector
aiosignal==1.4.0
    # via aiohttp
alembic==1.18.4
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
asn1crypto==1.5.1
    # via scramp
attrs==26.1.0
    # via aiohttp
authlib==1.6.9
    # via -r requirements.in
beautifulsoup4==4.14.3
    # via -r requirements.in
brotli==1.2.0
    # via -r requirements.in
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via uvicorn
cloud-sql-python-connector[pg8000]==1.20.1
    # via -r requirements.in
cryptography==46.0.5
    # via
    #   authlib
    #   cloud-sql-python-connector
    #   google-auth
dnspython==2.8.0
    # via cloud-sql-python-connector
fastapi==0.115.2
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.30.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.49.1
    # via
    #   -r requirements.in
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==3.10.0
    # via -r requirements.in
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.73.0
    # via google-api-core
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
pg8000==1.31.5
    # via cloud-sql-python-connector
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.1
    # via google-api-core
protobuf==6.33.6
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg[binary]==3.3.3
    # via -r requirements.in
psycopg-binary==3.3.3
    # via psycopg
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via pg8000
requests==2.32.5
    # via
    #   -r requirements.in
    #   cloud-sql-python-connector
    #   google-api-core
    #   google-cloud-storage
scramp==1.4.8
    # via pg8000
six==1.17.0
    # via python-dateutil
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.48
    # via
    #   -r requirements.in
    #   alembic
starlette==0.37.2
    # via
    #   -r requirements.in
    #   fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   beautifulsoup4
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
uvicorn==0.42.0
    # via -r requirements.in
yarl==1.23.0
    # via aiohttp
